@@ -30,16 +30,16 @@
     "prepublishOnly": "yarn clean && yarn build"
   },
   "devDependencies": {
-    "@types/react": "^17.0.13",
-    "@types/react-router-dom": "^5.1.7",
-    "react": "^17.0.2"
+    "@types/react": "^18",
+    "@types/react-router-dom": "^5.3.3",
+    "react": "^18"
   },
   "peerDependencies": {
-    "react": "^16.9.0 || ^17.0.0"
+    "react": "^16.9.0 || ^17.0.0 || ^18"
   },
   "dependencies": {
     "@elastic/eui": "^96.0.0",
-    "@mozartspa/mobx-form": "^0.6.0",
+    "@mozartspa/mobx-form": "^0.7.0-2",
     "@types/dlv": "^1.1.2",
     "@types/inflection": "^1.5.28",
     "@types/lodash.debounce": "^4.0.6",
@@ -50,12 +50,12 @@
     "lodash.debounce": "^4.0.8",
     "lodash.merge": "^4.6.2",
     "mobx": "^6.3.2",
-    "mobx-react-lite": "^3.2.0",
+    "mobx-react-lite": "^3.4.3",
     "node-polyglot": "^2.4.0",
     "query-string": "^7.0.1",
     "react-fast-compare": "^3.2.0",
-    "react-query": "^3.18.1",
-    "react-router-dom": "^5.2.0",
+    "react-query": "^3.39.3",
+    "react-router-dom": "^5.3.4",
     "rooks": "^5.0.2"
   }
 }
