@@ -28,23 +28,23 @@
     "prepublishOnly": "yarn clean && yarn build"
   },
   "devDependencies": {
-    "@mozartspa/mobx-form": "^0.6.0",
+    "@mozartspa/mobx-form": "^0.7.0-2",
     "@react-mool/core": "^0.12.0",
     "@rollup/plugin-commonjs": "^19.0.1",
     "@rollup/plugin-json": "^4.1.0",
     "@rollup/plugin-node-resolve": "^13.0.1",
     "@types/dlv": "^1.1.2",
-    "@types/react": "^17.0.13",
-    "@types/react-dom": "^17.0.8",
+    "@types/react": "^18",
+    "@types/react-dom": "^18",
     "dlv": "^1.1.3",
     "inflection": "^1.13.1",
     "mobx": "^6.3.2",
-    "mobx-react-lite": "^3.2.0",
+    "mobx-react-lite": "^3.4.3",
     "postcss": "^8.3.5",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router": "^5.2.0",
-    "react-router-dom": "^5.2.0",
+    "react": "^18",
+    "react-dom": "^18",
+    "react-router": "^5.3.4",
+    "react-router-dom": "^5.3.4",
     "rollup": "^2.53.2",
     "rollup-plugin-copy": "^3.4.0",
     "rollup-plugin-peer-deps-external": "^2.2.4",
@@ -54,7 +54,7 @@
   },
   "peerDependencies": {
     "@react-mool/core": "^0.1.0",
-    "react": "^16.9.0 || ^17.0.0"
+    "react": "^16.9.0 || ^17.0.0 || ^18"
   },
   "dependencies": {
     "@elastic/datemath": "^5.0.3",
@@ -63,7 +63,7 @@
     "@types/smoothscroll-polyfill": "^0.3.1",
     "moment": "^2.29.4",
     "prop-types": "^15.8.1",
-    "react-confirm": "^0.1.24",
+    "react-confirm": "^0.4.0",
     "react-error-boundary": "^3.1.3",
     "react-fast-compare": "^3.2.0",
     "smoothscroll-polyfill": "^0.4.4"
