@@ -18,12 +18,12 @@
     "@types/jest": "^26.0.15",
     "@types/node": "^12.0.0",
     "@types/pluralize": "^0.0.29",
-    "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
+    "@types/react": "^18",
+    "@types/react-dom": "^18",
     "json-graphql-server": "^2.2.3",
     "pluralize": "^8.0.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18",
+    "react-dom": "^18",
     "react-scripts": "^4.0.3",
     "typescript": "5.1.3",
     "web-vitals": "^1.0.1"
