@@ -1,13 +1,13 @@
 import React from "react"
-import ReactDOM from "react-dom"
+import { createRoot } from "react-dom/client"
 import App from "./App"
 import reportWebVitals from "./reportWebVitals"
 
-ReactDOM.render(
+const root = createRoot(document.getElementById("root") as HTMLElement)
+root.render(
   <React.StrictMode>
     <App />
-  </React.StrictMode>,
-  document.getElementById("root")
+  </React.StrictMode>
 )
 
 // If you want to start measuring performance in your app, pass a function
