@@ -6,7 +6,12 @@ export type TextAreaInputProps = InputProps & EuiTextAreaProps
 export const TextAreaInput = (props: TextAreaInputProps) => (
   <Input {...props}>
     {(field, inputProps) => (
-      <EuiTextArea {...field.input} fullWidth={props.fullWidth} {...inputProps} />
+      <EuiTextArea
+        {...field.input}
+        fullWidth={props.fullWidth}
+        {...inputProps}
+        children={undefined}
+      />
     )}
   </Input>
 )
